Stop using dotenv
It doesn't provide us any benefits in development (modifying .env forces
us to keep ignoring the changes in the working directory) or in
npm-/yarn-installed version (*where* would the end-user find the .env
file to modify?).

It's both enouth and easier for the user to provide the environment
variable directly (as the readme already documents), fewer moving pieces.