Document what OSes do we support and how
We've made some attempts to keep this package Windows-compatible in all
cases (like using rimraf instead of rm -rf to remove directories) but
this has proven untenable – not a single person in the company develops
on Windows and it's easy to introduce Unix-only development-only
incompatibilities. Case in point, there are two package.json scripts
that won't work on Windows:

    "test": "echo \"Error: no test specified\" && exit 1",
    "start": "npm run build && node build/index.js",

A good middle ground is to keep the package distributed to NPM
Windows-compatible (which is doable) but have no such requirements for
code that runs only during development and running the code directly.